+++ doc/conf.py
@@ -56,7 +56,9 @@ extensions = [
 smv_branch_whitelist = r'^(?!refs/heads/).*$'
 # v0.0.1 config is not compatible with sphinx-multiversion, so use 2 onwards
 # smv_tag_whitelist = r'^v\d+\.\d+.\d+$'
-smv_tag_whitelist = None
+# They say to set this to None, but then Sphinx complains about it not being
+# a string, so let's just use a regex that should lead to no tags
+smv_tag_whitelist = 'ignore all tags'
 # Mark vX.Y.Z as releases
 smv_released_pattern = r'^.*v.*$'
 
@@ -100,6 +102,8 @@ pygments_style = 'sphinx'
 
 # NumPyDoc configuration -----------------------------------------------------
 
+numpydoc_class_members_toctree = False
+numpydoc_attributes_as_param_list = True
 numpydoc_validate = True
 numpydoc_validation_checks = {'all'} | set(error_ignores)
 numpydoc_validation_exclude = {  # set of regex