+++ doc/changelog.rst
@@ -27,7 +27,7 @@ Enhancements
 
 * Add :meth:`mne_nirs.statistics.RegressionResults.save`. By `Robert Luke`_.
 * Add :meth:`mne_nirs.statistics.ContrastResults.save`. By `Robert Luke`_.
-* Add :func:`mne_nirs.statistics.load_glm`. By `Robert Luke`_.
+* Add :func:`mne_nirs.statistics.read_glm`. By `Robert Luke`_.
 
 
 Fixes